Obfuscated OpenSSH as a patch
The first import (commit 3a68b4ddd20b7bc1bedb0ed2bb0bfee9718f058c) refers to the files
as "Pristine OpenSSH 5.2 portable sources". Assuming that is correct, I checked out a second
copy of the repo to that particular commit and then did a

diff -Naur /path/to/pristine /path/to/obfuscated > obfuscated-openssh.patch

(the .git folders need to be moved away for this to work)

The output is 1,211 lines long.

Another command I considered was

git format-patch -12 HEAD --stdout

which produces a nice file, too, but each commit is applied in order (and the file is 1,630 lines long).